(PC-26529)[PRO] refactor: use slice for offers reducer
