+++ pro/src/store/offers/selectors.ts
@@ -1,4 +1,6 @@
 import { SearchFiltersParams } from 'core/Offers/types'
+import { RootState } from 'store/reducers'
 
-export const searchFiltersSelector = (state: any): SearchFiltersParams =>
-  state.offers.searchFilters
+export const searchFiltersSelector = (
+  state: RootState
+): Partial<SearchFiltersParams> => state.offers.searchFilters
